[Modify] handle event
