+++ src/layouts/Condition/index.jsx
@@ -16,7 +16,15 @@ import ListItemText from '@material-ui/core/ListItemText';
 import EditIcon from '@material-ui/icons/Edit';
 import _ from 'lodash';
 import DeleteIcon from '@material-ui/icons/Delete'
-import ConditionDialog from 'components/ConditionDialog'
+import Typography from '@material-ui/core/Typography';
+import CloseIcon from '@material-ui/icons/Close';
+import Dialog from '@material-ui/core/Dialog';
+import DialogActions from '@material-ui/core/DialogActions';
+import DialogContent from '@material-ui/core/DialogContent';
+import DialogTitle from '@material-ui/core/DialogTitle';
+import TextField from '@material-ui/core/TextField';
+import Checkbox from '@material-ui/core/Checkbox';
+import MenuItem from '@material-ui/core/MenuItem';
 
 const styles = theme => ({
   card: {
@@ -46,7 +54,13 @@ const theme = createMuiTheme({
   },
 });
 
+// Menu Item
 const conditionMenu = ['Moving', 'Riding bike', 'Static'];
+const hourInterval = Array(12).fill().map((value, index) => index + 1);
+const minInterval = Array(60).fill().map((value, index) => index + 1);
+const amPm = ['AM', 'PM'];
+const timeUnit = ['hour', 'minute', 'second'];
+const optionMenu = ['weak light', 'lower noise', 'battery', 'speed', 'GPS', 'location'];
 
 class Condition extends React.Component {
 
@@ -132,8 +146,16 @@ class Condition extends React.Component {
     this.setState({conditionList: tmp});
   };
 
-  handleClose = () => {
-    this.setState({ open: false });
+  handleClose = (index) => {
+    let tmp = this.state.conditionList;
+    tmp[index].isOpen = false;
+    this.setState({conditionList: tmp});
+  };
+
+  handleChange = (event) => {
+    let tmp = this.state.conditionList;
+    tmp[0][event.target.id] = event.target.value;
+    this.setState({conditionList: tmp});
   };
 
   render() {
@@ -160,9 +182,269 @@ class Condition extends React.Component {
                     <DeleteIcon />
                   </IconButton>
                 </ListItemSecondaryAction>
-                <ConditionDialog
-                  contentObj = {condition}
-                />
+              <Dialog
+                open={condition.isOpen}
+                onClose={(e) => this.handleClose(index)}
+              >
+                <DialogTitle>
+                  <TextField
+                    id="name"
+                    label="Name"
+                    className={classes.textField}
+                    value={condition.name}
+                    index={index}
+                    onChange={this.handleChange}
+                    margin="normal"
+                  />
+                </DialogTitle>
+                <DialogContent>
+                  <List>
+                    <ListItem
+                      role={undefined}
+                      dense
+                      className={classes.listItem}
+                    >
+                      <Checkbox
+                        tabIndex={-1}
+                        disableRipple
+                      />
+                      <ListItemText primary='From' />
+                      <TextField
+                        id=""
+                        select
+                        className={classes.textField}
+                        value='10'
+                        SelectProps={{
+                          MenuProps: {
+                            className: classes.menu,
+                          },
+                        }}
+                        margin="normal"
+                      >
+                        {hourInterval.map(option => (
+                          <MenuItem key={option} value={option}>
+                            {option}
+                          </MenuItem>
+                        ))}
+                      </TextField>
+                      <ListItemText primary='：' />
+                      <TextField
+                        id="select-currency"
+                        select
+                        className={classes.textField}
+                        onChange={this.handleChange('name')}
+                        SelectProps={{
+                          MenuProps: {
+                            className: classes.menu,
+                          },
+                        }}
+                        margin="normal"
+                      >
+                        {minInterval.map(option => (
+                          <MenuItem key={option} value={option}>
+                            {option}
+                          </MenuItem>
+                        ))}
+                      </TextField>
+                      <ListItemText primary=' ' />
+                      <TextField
+                        id="select-currency"
+                        select
+                        className={classes.textField}
+                      
+                        onChange={this.handleChange('name')}
+                        SelectProps={{
+                          MenuProps: {
+                            className: classes.menu,
+                          },
+                        }}
+                        margin="normal"
+                      >
+                        {amPm.map(option => (
+                          <MenuItem key={option} value={option}>
+                            {option}
+                          </MenuItem>
+                        ))}
+                      </TextField>
+                      <ListItemText primary='to' />
+                      <TextField
+                        id="select-currency"
+                        select
+                        className={classes.textField}
+                        v
+                        onChange={this.handleChange('name')}
+                        SelectProps={{
+                          MenuProps: {
+                            className: classes.menu,
+                          },
+                        }}
+                        margin="normal"
+                      >
+                        {hourInterval.map(option => (
+                          <MenuItem key={option} value={option}>
+                            {option}
+                          </MenuItem>
+                        ))}
+                      </TextField>
+                      <ListItemText primary='：' />
+                      <TextField
+                        id="select-currency"
+                        select
+                        className={classes.textField}
+                        
+                        onChange={this.handleChange('currency')}
+                        SelectProps={{
+                          MenuProps: {
+                            className: classes.menu,
+                          },
+                        }}
+                        margin="normal"
+                      >
+                        {minInterval.map(option => (
+                          <MenuItem key={option} value={option}>
+                            {option}
+                          </MenuItem>
+                        ))}
+                      </TextField>
+                      <ListItemText primary=' ' />
+                      <TextField
+                        id="select-currency"
+                        select
+                        className={classes.textField}
+                        
+                        onChange={this.handleChange('currency')}
+                        SelectProps={{
+                          MenuProps: {
+                            className: classes.menu,
+                          },
+                        }}
+                        margin="normal"
+                      >
+                        {amPm.map(option => (
+                          <MenuItem key={option} value={option}>
+                            {option}
+                          </MenuItem>
+                        ))}
+                      </TextField>
+                    </ListItem>
+                    <ListItem
+                      role={undefined}
+                      dense
+                      className={classes.listItem}
+                    >
+                      <Checkbox
+                        tabIndex={-1}
+                        disableRipple
+                      />
+                      <ListItemText primary='Last for' />
+                      <TextField
+                        id="select-currency"
+                        select
+                        className={classes.textField}
+                        
+                        onChange={this.handleChange('currency')}
+                        SelectProps={{
+                          MenuProps: {
+                            className: classes.menu,
+                          },
+                        }}
+                        margin="normal"
+                      >
+                        {hourInterval.map(option => (
+                          <MenuItem key={option} value={option}>
+                            {option}
+                          </MenuItem>
+                        ))}
+                      </TextField>
+                      <ListItemText primary='：' />
+                      <TextField
+                        id="select-currency"
+                        select
+                        className={classes.textField}
+                        
+                        onChange={this.handleChange('currency')}
+                        SelectProps={{
+                          MenuProps: {
+                            className: classes.menu,
+                          },
+                        }}
+                        margin="normal"
+                      >
+                        {minInterval.map(option => (
+                          <MenuItem key={option} value={option}>
+                            {option}
+                          </MenuItem>
+                        ))}
+                      </TextField>
+                      <ListItemText primary=' ' />
+                      <TextField
+                        id="select-currency"
+                        select
+                        className={classes.textField}
+                        
+                        onChange={this.handleChange('currency')}
+                        SelectProps={{
+                          MenuProps: {
+                            className: classes.menu,
+                          },
+                        }}
+                        margin="normal"
+                      >
+                        {timeUnit.map(option => (
+                          <MenuItem key={option} value={option}>
+                            {option}
+                          </MenuItem>
+                        ))}
+                      </TextField>
+                    </ListItem>
+                    <ListItem
+                      divider = {true}
+                    >
+                      <Typography variant="subheading">
+                        Rule
+                      </Typography>
+                    </ListItem>
+                    <ListItem>
+                      <TextField
+                          id=""
+                          select
+                          className={classes.textField}
+                          SelectProps={{
+                            MenuProps: {
+                              className: classes.menu,
+                            },
+                          }}
+                          margin="normal"
+                        >
+                          {optionMenu.map(option => (
+                            <MenuItem key={option} value={option}>
+                              {option}
+                            </MenuItem>
+                          ))}
+                      </TextField>
+                      <IconButton>
+                          <CloseIcon />
+                      </IconButton>
+                    </ListItem>
+                    <ListItem>
+                      <Button 
+                        variant="contained" 
+                        color="primary" 
+                      >
+                        + rule
+                      </Button>
+                    </ListItem>
+                  </List>
+                </DialogContent>
+                <DialogActions>
+                  <Button onClick={this.handleClose} color="primary">
+                    cancel
+                  </Button>
+                  <Button onClick={this.handleClose} color="primary" autoFocus>
+                    done
+                  </Button>
+                </DialogActions>
+              </Dialog>
               </ListItem>
               )
             }
@@ -179,7 +461,6 @@ class Condition extends React.Component {
             </Button>
           </CardActions>
         </Card>
-        
         </MuiThemeProvider>
       </div>
     );
